services/aivo-brain-svc/requirements.txt
# AIVO Main Brain Service - ACTUAL AI DEPENDENCIES
# This is the REAL foundation model service - not a mock!

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# ML/AI - PyTorch with CUDA support
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0

# Transformers and LLM libraries
transformers==4.36.0
accelerate==0.25.0
bitsandbytes==0.41.3
peft==0.7.0
sentencepiece==0.1.99
protobuf==3.20.3

# Cloud AI providers
openai==1.58.1
anthropic==0.72.1
google-generativeai==0.8.3

# Model optimization
optimum==1.14.1
onnx==1.15.0
onnxruntime==1.16.3
onnxruntime-gpu==1.16.3

# Data processing
datasets==2.15.0
numpy==1.24.3
pandas==2.1.4
scipy==1.11.4

# Caching and storage
redis==5.0.1
boto3==1.34.0
minio==7.2.0

# Monitoring and logging
prometheus-client==0.19.0
python-json-logger==2.0.7
sentry-sdk==1.40.0

# Resilience and fault tolerance
circuitbreaker==1.4.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pylint==3.2.7
